refactor: map controller context 로직 코드 포맷 통일 #12
