+++ src/app/layout.tsx
@@ -3,8 +3,8 @@ import { Noto_Sans_KR } from 'next/font/google';
 import NavigationBar from '@/components/NavigationBar';
 import ReactQueryProvider from '@/components/common/ReactQueryProvider';
 import SearchInputField from '@/components/common/SearchInputField';
+import MapControllerProvider from '@/components/contexts/MapControllerProvider';
 import PreferencesProvider from '@/components/contexts/PreferencesProvider';
-import MapControllerProvider from '@/components/maps/MapControllerProvider';
 import MapView from '@/components/maps/MapView';
 import SearchedStoreList from '@/components/stores/SearchedStoreList';
 import './globals.css';
